requirements.txt
aiohttp==3.9.3
quart==0.19.4
quart-cors==0.7.0
azure-storage-blob==12.19.1
azure-identity==1.15.0
redis